requirements.txt
Flask==2.0.2
Flask_Compress==1.10.1
furl==2.1.3
python-dotenv==0.19.2
requests==2.26.0
